fix(storage): add integration test entry point for storage tests (#64)
