+++ minigu/storage/tests/integration_tests.rs
@@ -1,0 +1,5 @@
+// Integration tests for minigu-storage
+
+mod ap;
+mod common;
+mod tp;
