+++ minigu/storage/tests/tp/isolation_level_test.rs
@@ -1,12 +1,12 @@
-mod common;
 use std::thread;
 
-use common::*;
 use minigu_common::value::ScalarValue;
-use minigu_storage::IsolationLevel;
 use minigu_storage::model::edge::Edge;
 use minigu_storage::model::properties::PropertyRecord;
 use minigu_storage::model::vertex::Vertex;
+use minigu_storage::tp::IsolationLevel;
+
+use crate::common::*;
 
 // ========== DIRTY READ TESTS ==========
